throttle visits and sequence refreshes independently
